backend/requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
pymongo
langchain
langgraph
line-bot-sdk       # LINE Bot 用
beautifulsoup4     # 爬蟲（選用）
